requirements-test.txt
# Home Assistant testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-timeout>=2.1.0
pytest-sugar>=0.9.7

# Home Assistant test utilities
homeassistant>=2023.1.0
homeassistant[test]>=2023.1.0

# Mock and testing utilities
requests-mock>=1.10.0
syrupy>=4.0.0
freezegun>=1.2.0

# Code quality and additional testing tools
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
pre-commit>=3.0.0

# Additional testing dependencies for async and HTTP testing
aioresponses>=0.7.4
pytest-aiohttp>=1.0.4